Finished archive page
Finished the Live Stream archive page. Each video has a private Vimeo
link that can only be access through this site. Videos.yml gets updated
with the new info and the layout builds the page, cycling over each
entry in the YAML page and displaying it with the proper private Vimeo
Link.